refac(campagne) - Extract config in new service
