+++ src/jdRoll/controller/public_campagn.php
@@ -53,7 +53,7 @@ $publicCampagneController->get('/{id}', function($id) use($app) {
 })->bind("campagne");
 
 $publicCampagneController->get('/config/{id}', function($id) use($app) {
-	$campagne = $app['campagneService']->getCampagneConfig($id);
+	$campagne = $app["campagneConfigService"]->getCampagneConfig($id);
 	return $app->render('game/config/css.html.twig', ['campagne_config' => $campagne, 'error' => ""]);
 })->bind("campagne_config");
